MWPW-118484 - Rename analytics
* Moved martech into martech folder
* Moved and renamed analytics.js to attributes.js

Resolves: [MWPW-118484](https://jira.corp.adobe.com/browse/MWPW-118484)
